requirements.txt
Flask==2.2.5   
Werkzeug==2.2.3
openai==0.27.2
gunicorn==20.1.0
python-dotenv==0.21.0
requests==2.28.2
